+++ mrfmsim/component/__init__.py
@@ -1,0 +1,4 @@
+from .magnet import SphereMagnet, RectangularMagnet
+from .cantilever import Cantilever
+from .grid import Grid
+from .sample import Sample